fix: split tech catalog file under repo non-comment line limit
Move late catalog chunk helpers into a Dart part file so tech_catalog.dart stays within the repo 1000 non-comment-line convention while preserving the same tech definitions and build order.

Made-with: Cursor
